Move some classes to LegoOmni
