+++ LEGO1/lego/legoomni/src/common/mxcompositemediapresenter.cpp
