+++ LEGO1/lego/legoomni/src/video/mxtransitionmanager.cpp
